% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substr_by_cha.R
\name{substr_by_cha}
\alias{substr_by_cha}
\title{Divide character strings given a certain character.}
\usage{
substr_by_cha(cha_str, delim = "|", type = c("before", "after"))
}
\arguments{
\item{cha_str}{A character string or a vector of character string.}

\item{delim}{Delimiter character. Default is "|".}

\item{type}{Return the partial character string before or after the delimiter.}
}
\value{
Return the desired substring (an individual or a vector).
}
\description{
This is a handy function to divide character strings given a certain character. This is especially useful
when sometimes gene names show up in the form of "GENESYMBOL|ENTREZID"
}
\examples{
substr_by_cha("ABCB1|5243", delim = "|", type = "before")
substr_by_cha(c("ABCB1-5243", "ABL1-25", "ABL2-27"), delim = "-", type = "after")
}
\keyword{misc}
\keyword{utilities}
